Be explicit here and cast it
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@62304 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/gd/gd.c
@@ -3534,8 +3534,8 @@ static void _php_image_convert(INTERNAL_FUNCTION_PARAMETERS, int image_type )
 			else {
 				y_ratio = x_ratio;
 			}
-			dest_width = org_width / x_ratio;
-			dest_height = org_height / y_ratio;
+			dest_width = (int)(org_width / x_ratio);
+			dest_height = (int)(org_height / y_ratio);
 		}
 		else {
 			x_ratio = (float) dest_width / (float) org_width;
@@ -3547,8 +3547,8 @@ static void _php_image_convert(INTERNAL_FUNCTION_PARAMETERS, int image_type )
 			else {
 				y_ratio = x_ratio;
 			}
-			dest_width = org_width * x_ratio;
-			dest_height = org_height * y_ratio;
+			dest_width = (int)(org_width * x_ratio);
+			dest_height = (int)(org_height * y_ratio);
 		}
 
 		im_tmp = gdImageCreate (dest_width, dest_height);
